Change all int to long for id in Migrations

+++ Flower_shop/Controllers/AdminPlatformController.cs
@@ -125,7 +125,7 @@ namespace Flower_shop.Controllers
             return View();
         }
         [HttpPost]
-        public async Task<IActionResult> ProductDelete(int id)
+        public async Task<IActionResult> ProductDelete(long id)
         {
             await _productRepository.RemoveByIdAsync(id);
 
@@ -137,7 +137,7 @@ namespace Flower_shop.Controllers
             return View();
         }
         [HttpPost]
-        public async Task<IActionResult> TypeProductDelete(int typeId)
+        public async Task<IActionResult> TypeProductDelete(long typeId)
         {
             await _typeProductRepository.RemoveTypeProductAsync(typeId);
